+++ tests/ConfigTest.php
@@ -7,6 +7,7 @@ namespace FormatPHP\Test;
 use FormatPHP\Config;
 use FormatPHP\Extractor\IdInterpolator;
 use FormatPHP\Intl\Locale;
+use Locale as PhpLocale;
 
 class ConfigTest extends TestCase
 {
@@ -18,12 +19,24 @@ class ConfigTest extends TestCase
             'em' => fn (string $text): string => '<em class="bar">' . $text . '</em>',
             'strong' => fn (string $text): string => '<strong class="foo">' . $text . '</strong>',
         ];
+        $idInterpolatorPattern = '[md5:contenthash:base64:16]';
 
-        $config = new Config($locale, $defaultLocale, $defaultRichTextElements);
+        $config = new Config($locale, $defaultLocale, $defaultRichTextElements, $idInterpolatorPattern);
 
         $this->assertSame($locale, $config->getLocale());
         $this->assertSame($defaultLocale, $config->getDefaultLocale());
         $this->assertSame($defaultRichTextElements, $config->getDefaultRichTextElements());
+        $this->assertSame($idInterpolatorPattern, $config->getIdInterpolatorPattern());
+    }
+
+    public function testConfigDefaults(): void
+    {
+        $systemLocale = new Locale(PhpLocale::getDefault());
+        $config = new Config();
+
+        $this->assertSame($systemLocale->toString(), $config->getLocale()->toString());
+        $this->assertNull($config->getDefaultLocale());
+        $this->assertSame([], $config->getDefaultRichTextElements());
         $this->assertSame(IdInterpolator::DEFAULT_ID_INTERPOLATION_PATTERN, $config->getIdInterpolatorPattern());
     }
 }